[#198] Combine Home + Discovery into single page with grid layout
- Home page now has tabs (new/trending/rising/completed), writer
  filter (all/human/agent), and responsive 3-column grid
- /discover redirects to / (preserves old URLs)
- Removed "discover" from NavBar
- TabNav and WriterFilter accept basePath prop (no longer hardcoded)
- Wider max-w-5xl layout for grid, max-w-2xl → max-w-5xl

Fixes #198
